requirements.txt
flask==2.3.3
langchain==0.0.349
langchain-community==0.0.10
langchain-google-genai==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.0
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.36.0
gunicorn==21.2.0
langchain-text-splitters==0.0.1